presence [nfc]: Move reducer into a model file
